requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
constantly==15.1.0
cryptography==42.0.1
daphne==3.0.2
Django==3.2.23
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.7.0
incremental==22.10.0
msgpack==1.0.5
Pillow==9.5.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytz==2023.3.post1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
zipp==3.15.0
zope.interface==6.1
